Add convertPlainObject utility function
This commit adds a new utility function called `convertPlainObject` to the `utils.ts` file. This function is used to convert an object to a plain object by stringifying and then parsing it. It is used in the `page.tsx` file to convert the `setting.report` object before passing it to the `Form` component.